+++ datagrowth/resources/shell/generic.py
@@ -9,10 +9,7 @@ from time import sleep
 from django.db import models
 from django.core.exceptions import ValidationError
 from django.utils.timezone import now
-try:
-    from django.db.models import JSONField
-except ImportError:
-    from django.contrib.postgres.fields import JSONField
+from django.db.models import JSONField
 
 from datagrowth import settings as datagrowth_settings
 from datagrowth.resources.base import Resource