+++ tests/datatypes/models/document.py
@@ -1,8 +1,3 @@
-try:
-    from django.db.models import JSONField
-except ImportError:
-    from django.contrib.postgres.fields import JSONField
-
 from datagrowth.datatypes import DocumentBase